Bugfix: make distance() work with no pressure argument
This is a bit subtle: we want an immutable default, p=0,
but internally p needs to be an array.  When p is given
explicitly as a scalar argument, the match_args_return
decorator takes care of it, but to allow it to be optional
we need to do the conversion to array in the function
itself.